# docs/swagger/manager/manager_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing shards and cluster state",
        "title": "Sharding System Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@sharding-system.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/reshard/jobs/{id}": {
            "get": {
                "description": "Retrieves the status of a resharding job by job ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resharding"
                ],
                "summary": "Get reshard job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job status",
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.ReshardJob"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/reshard/merge": {
            "post": {
                "description": "Merges multiple source shards into a target shard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resharding"
                ],
                "summary": "Merge shards",
                "parameters": [
                    {
                        "description": "Merge Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.MergeRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Merge job started",
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.ReshardJob"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/reshard/split": {
            "post": {
                "description": "Splits a shard into multiple target shards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resharding"
                ],
                "summary": "Split a shard",
                "parameters": [
                    {
                        "description": "Split Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.SplitRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Split job started",
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.ReshardJob"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/shards": {
            "get": {
                "description": "Returns a list of all shards in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "List all shards",
                "responses": {
                    "200": {
                        "description": "List of shards",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sharding-system_pkg_models.Shard"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new database shard with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Create a new shard",
                "parameters": [
                    {
                        "description": "Shard Configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.CreateShardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Shard created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.Shard"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/shards/{id}": {
            "get": {
                "description": "Retrieves shard information by shard ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Get shard by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shard information",
                        "schema": {
                            "$ref": "#/definitions/github_com_sharding-system_pkg_models.Shard"
                        }
                    },
                    "404": {
                        "description": "Shard not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a shard by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Delete a shard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Shard deleted successfully"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/shards/{id}/promote": {
            "post": {
                "description": "Promotes a replica to become the primary shard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shards"
                ],
                "summary": "Promote a replica to primary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Replica Promotion Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Replica promoted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_sharding-system_pkg_models.MergeRequest": {
            "type": "object",
            "properties": {
                "source_shard_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "target_shard": {
                    "$ref": "#/definitions/github_com_sharding-system_pkg_models.CreateShardRequest"
                }
            }
        },
...
        "github_com_sharding-system_pkg_models.ReshardJob": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "keys_migrated": {
                    "type": "integer"
                },
                "progress": {
                    "description": "0.0 to 1.0",
                    "type": "number"
                },
                "source_shards": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "description": "\"pending\", \"precopy\", \"deltasync\", \"cutover\", \"completed\", \"failed\"",
                    "type": "string"
                },
                "target_shards": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_keys": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"split\" or \"merge\"",
                    "type": "string"
                }
            }
        },
        "github_com_sharding-system_pkg_models.Shard": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "hash_range_end": {
                    "type": "integer"
                },
                "hash_range_start": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primary_endpoint": {
                    "type": "string"
                },
                "replicas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "\"active\", \"migrating\", \"readonly\", \"inactive\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                },
                "vnodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sharding-system_pkg_models.VNode"
                    }
                }
            }
        },
        "github_com_sharding-system_pkg_models.SplitRequest": {
            "type": "object",
            "properties": {
                "source_shard_id": {
                    "type": "string"
                },
                "split_point": {
                    "description": "Optional explicit split point",
                    "type": "integer"
                },
                "target_shards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sharding-system_pkg_models.CreateShardRequest"
                    }
                }
            }
        },
...
    }
}